+++ packages/chains/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wagmi/chains
 
+## 0.1.2
+
+### Patch Changes
+
+- 36a2131: Added Taraxa Chains.
+
+## 0.1.1
+
+### Patch Changes
+
+- ccf521c: Fixed Fantom Testnet Chain ID.
+
 ## 0.1.0
 
 ### Minor Changes